+++ modules/schematics-core/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@ngrx/schematics-core",
+  "version": "0.0.0-PLACEHOLDER",
+  "description": "NgRx Schematics Utilities",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ngrx/platform.git"
+  },
+  "keywords": [
+    "RxJS",
+    "Angular",
+    "Redux",
+    "NgRx",
+    "Schematics",
+    "Angular CLI"
+  ],
+  "author": "Brandon Roberts <robertsbt@gmail.com>",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/ngrx/platform/issues"
+  },
+  "homepage": "https://github.com/ngrx/platform#readme",
+  "peerDependencies": {
+    "@angular-devkit/core": "^0.5.0",
+    "@angular-devkit/schematics": "^0.5.0"
+  }
+}
