+++ modules/schematics-core/testing/get-file-content.ts
