+++ modules/schematics-core/testing/index.ts
@@ -1,2 +1,4 @@
 export * from './create-app-module';
+export * from './create-reducers';
+export * from './create-workspace';
 export * from './get-file-content';
