+++ modules/schematics-core/utility/strings.ts
