+++ modules/schematics/src/action/index.ts
@@ -14,9 +14,8 @@ import {
   Tree,
   SchematicContext,
 } from '@angular-devkit/schematics';
-import * as stringUtils from '../strings';
 import { Schema as ActionOptions } from './schema';
-import { getProjectPath } from '../utility/project';
+import { getProjectPath, stringUtils } from '@ngrx/schematics-core';
 
 export default function(options: ActionOptions): Rule {
   return (host: Tree, context: SchematicContext) => {
