+++ modules/schematics/src/cli.spec.ts
@@ -5,9 +5,9 @@ import {
 } from '@angular-devkit/schematics/testing';
 import * as path from 'path';
 import {
+  getTestProjectPath,
   createWorkspace,
-  getProjectPath,
-} from './utility/test/create-workspace';
+} from '@ngrx/schematics-core/testing';
 
 describe('CLI Schematic', () => {
   const schematicRunner = new SchematicTestRunner(
@@ -20,7 +20,7 @@ describe('CLI Schematic', () => {
     project: 'bar',
   };
 
-  const projectPath = getProjectPath();
+  const projectPath = getTestProjectPath();
 
   let appTree: UnitTestTree;
