+++ modules/schematics/src/entity/index.ts
@@ -14,14 +14,14 @@ import {
   Tree,
   SchematicContext,
 } from '@angular-devkit/schematics';
-import * as stringUtils from '../strings';
-import { Schema as EntityOptions } from './schema';
 import {
+  stringUtils,
   addReducerToState,
   addReducerImportToNgModule,
-} from '../utility/ngrx-utils';
-import { findModuleFromOptions } from '../utility/find-module';
-import { getProjectPath } from '../utility/project';
+  getProjectPath,
+  findModuleFromOptions,
+} from '@ngrx/schematics-core';
+import { Schema as EntityOptions } from './schema';
 
 export default function(options: EntityOptions): Rule {
   return (host: Tree, context: SchematicContext) => {
