+++ modules/schematics/src/reducer/index.ts
@@ -15,16 +15,14 @@ import {
   url,
 } from '@angular-devkit/schematics';
 import * as ts from 'typescript';
-import * as stringUtils from '../strings';
-import { findModuleFromOptions } from '../utility/find-module';
-import { Schema as ReducerOptions } from './schema';
 import {
-  addReducerToStateInferface,
-  addReducerToActionReducerMap,
+  getProjectPath,
+  findModuleFromOptions,
+  stringUtils,
   addReducerToState,
   addReducerImportToNgModule,
-} from '../utility/ngrx-utils';
-import { getProjectPath } from '../utility/project';
+} from '@ngrx/schematics-core';
+import { Schema as ReducerOptions } from './schema';
 
 export default function(options: ReducerOptions): Rule {
   return (host: Tree, context: SchematicContext) => {
